Add PublicationDate and DateRegistration columns
This commit introduces two new columns: `PublicationDate` to the `Mods` table and `DateRegistration` to the `Creators` table. Both columns are of type `DateTime`, non-nullable, and have a default value of `new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified)`. The `DataContextModelSnapshot.cs` file is also updated to reflect these changes in the Entity Framework Core model.